refactor code to ad-helper
